Dealing with TODO within core module.
