+++ accounting/accounting/src/main/java/liquid/accounting/controller/ChargeController.java
@@ -12,8 +12,8 @@ import liquid.accounting.service.ExchangeRateService;
 import liquid.accounting.service.InternalChargeService;
 import liquid.container.domain.ContainerCap;
 import liquid.container.domain.ContainerType;
-import liquid.domain.LoadingType;
-import liquid.domain.TradeType;
+import liquid.order.domain.LoadingType;
+import liquid.order.domain.TradeType;
 import liquid.operation.domain.ServiceProvider;
 import liquid.operation.domain.ServiceSubtype;
 import liquid.operation.service.ServiceProviderService;
@@ -24,7 +24,7 @@ import liquid.process.service.TaskService;
 import liquid.transport.domain.LegEntity;
 import liquid.transport.domain.TransMode;
 import liquid.transport.service.LegService;
-import liquid.web.domain.ExchangeRateDto;
+import liquid.accounting.model.ExchangeRateDto;
 import liquid.web.domain.SearchBarForm;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
