+++ order/order-api/src/main/java/liquid/order/domain/LoadingType.java
@@ -1,20 +1,19 @@
-package liquid.domain;
+package liquid.order.domain;
 
 /**
- * TODO: Comments.
  * User: tao
  * Date: 9/29/13
- * Time: 10:24 PM
+ * Time: 10:29 PM
  */
-public enum TradeType {
-    DOMESTIC(0, "trade.domestic"),
-    FOREIGN(1, "trade.foreign");
+public enum LoadingType {
+    YARD(0, "loading.yard"),
+    TRUCK(1, "loading.truck");
 
     private final int type;
 
     private final String i18nKey;
 
-    private TradeType(int type, String i18nKey) {
+    private LoadingType(int type, String i18nKey) {
         this.type = type;
         this.i18nKey = i18nKey;
     }
@@ -27,12 +26,12 @@ public enum TradeType {
         return i18nKey;
     }
 
-    public static TradeType valueOf(int type) {
+    public static LoadingType valueOf(int type) {
         switch (type) {
             case 0:
-                return DOMESTIC;
+                return YARD;
             case 1:
-                return FOREIGN;
+                return TRUCK;
             default:
                 throw new IllegalArgumentException(String.format("%s should be from %s and %s.", type, 0, 1));
         }
@@ -41,7 +40,7 @@ public enum TradeType {
     @Override
     public String toString() {
         final StringBuilder sb = new StringBuilder(super.toString());
-        sb.append("TradeType{");
+        sb.append("LoadingType{");
         sb.append("type=").append(type);
         sb.append(", i18nKey='").append(i18nKey).append('\'');
         sb.append('}');
