+++ order/order-api/src/main/java/liquid/order/domain/TradeType.java
@@ -1,20 +1,19 @@
-package liquid.domain;
+package liquid.order.domain;
 
 /**
- * TODO: Comments.
  * User: tao
  * Date: 9/29/13
- * Time: 10:29 PM
+ * Time: 10:24 PM
  */
-public enum LoadingType {
-    YARD(0, "loading.yard"),
-    TRUCK(1, "loading.truck");
+public enum TradeType {
+    DOMESTIC(0, "trade.domestic"),
+    FOREIGN(1, "trade.foreign");
 
     private final int type;
 
     private final String i18nKey;
 
-    private LoadingType(int type, String i18nKey) {
+    private TradeType(int type, String i18nKey) {
         this.type = type;
         this.i18nKey = i18nKey;
     }
@@ -27,12 +26,12 @@ public enum LoadingType {
         return i18nKey;
     }
 
-    public static LoadingType valueOf(int type) {
+    public static TradeType valueOf(int type) {
         switch (type) {
             case 0:
-                return YARD;
+                return DOMESTIC;
             case 1:
-                return TRUCK;
+                return FOREIGN;
             default:
                 throw new IllegalArgumentException(String.format("%s should be from %s and %s.", type, 0, 1));
         }
@@ -41,7 +40,7 @@ public enum LoadingType {
     @Override
     public String toString() {
         final StringBuilder sb = new StringBuilder(super.toString());
-        sb.append("LoadingType{");
+        sb.append("TradeType{");
         sb.append("type=").append(type);
         sb.append(", i18nKey='").append(i18nKey).append('\'');
         sb.append('}');
