+++ order/order/src/main/java/liquid/order/controller/OrderController.java
@@ -14,8 +14,8 @@ import liquid.container.domain.ContainerCap;
 import liquid.container.domain.ContainerSubtypeEntity;
 import liquid.container.domain.ContainerType;
 import liquid.container.service.ContainerSubtypeService;
-import liquid.domain.LoadingType;
-import liquid.domain.TradeType;
+import liquid.order.domain.LoadingType;
+import liquid.order.domain.TradeType;
 import liquid.operation.domain.*;
 import liquid.operation.service.*;
 import liquid.order.domain.OrderEntity;
