+++ order/order/src/main/java/liquid/order/facade/OrderFacade.java
@@ -5,8 +5,8 @@ import liquid.accounting.model.ReceivableSummary;
 import liquid.container.domain.ContainerType;
 import liquid.container.service.ContainerSubtypeService;
 import liquid.domain.Currency;
-import liquid.domain.LoadingType;
-import liquid.domain.TradeType;
+import liquid.order.domain.LoadingType;
+import liquid.order.domain.TradeType;
 import liquid.operation.domain.Customer;
 import liquid.operation.domain.Location;
 import liquid.operation.domain.ServiceSubtype;
